# Optimized `diff()` function which now uses `fastDiff()` function internally for large data sets