--- a/src/diff.js
+++ b/src/diff.js
@@ -32,7 +32,7 @@ export default function diff( a, b, cmp ) {
 	const aLength = a.length;
 	const bLength = b.length;
 
-	// Perform `fastDiff` for longer strings/arrays.
+	// Perform `fastDiff` for longer strings/arrays (see #269).
 	if ( Math.min( aLength, bLength ) >= 400 && aLength + bLength >= 1300 || aLength + bLength >= 2000 ) {
 		return diff.fastDiff( a, b, cmp, true );
 	}