--- a/tests/manual/diff/diff.html
+++ b/tests/manual/diff/diff.html
@@ -1,0 +1,1 @@
+See browser dev console output.