--- a/tests/manual/diff/diff.md
+++ b/tests/manual/diff/diff.md
@@ -1,0 +1,3 @@
+## `diff()`
+
+Checks `diff()` performance.