… tests/_utils/longtext.txt
… tests/diff.js
… tests/fastdiff.js
… tests/manual/diff/diff.js