### Suggested merge commit message ([convention](https://github.com/ckeditor/ckeditor5-design/wiki/Git-commit-message-convention))

Other: Optimized `diff()` function which now uses `fastDiff()` function internally for large data sets. Closes #269.

---

### Additional information

See #269.