@f1ames 2019-02-11
I was thinking about adding some performance unit tests which checks how long setting data (with `editor.setData()`) takes. There is no sense in measuring how long `diff()` function takes, because it doesn't change very often and as you can see from https://github.com/ckeditor/ckeditor5-utils/issues/269#issuecomment-461116792, this issue was caused by some external code changes (not the `diff()` changes itself).

The problem with such tests is that exec time may vary (same test from 100 to 400 ms between runs in different browsers, or even 400 to 1100 ms) so I'm not sure if it makes sense as such tests are not very stable and the exec time depends on the environment they are run in.