@Reinmar 2019-02-11
> I was thinking about adding some performance unit tests which checks how long setting data (with `editor.setData()`) takes.

In general, 👍 .

> Anyway, these tests can be added separately (especially that they should be added to core IMHO) so this PR is ready for review.

Definitely to the core. 



> The problem with such tests is that exec time may vary (same test from 100 to 400 ms between runs in different browsers, or even 400 to 1100 ms) so I'm not sure if it makes sense as such tests are not very stable and the exec time depends on the environment they are run in. WDYT @Reinmar?

I assume we changed results from a scenario with e.g. 1000 nodes from 10s to 100ms, am I right? So it's easy to write a threshold value – anything below e.g. 1s will be fine. Such a test should be really stable.

However, tests should be fast. So ideally I'd love a test which always passes in <100ms. Is there such a node count which would generate such results now and before would fail?