@Reinmar 2019-02-11
> Definitely to the core.
> 

Actually... to the engine. The bug was in the engine.