@f1ames 2019-02-11
> I assume we changed results from a scenario with e.g. 1000 nodes from 10s to 100ms, am I right? So it's easy to write a threshold value – anything below e.g. 1s will be fine. Such a test should be really stable.

Yes and no :D Previously `render()` took around 23 seconds for the sample data, now it takes around 200 - 300ms. However, the other parts of the pipeline still takes visible amount of time

![image](https://user-images.githubusercontent.com/1061942/52574330-54e88f80-2e1c-11e9-802d-7452195674ad.png)

which means end to end tests for larger data sets still takes up to few seconds to be completed. Maybe we could test only some part of the pipeline (e.g. `render()`?), but the whole picture might get lost then.