@Mgsy 2019-02-12
## Steps to reproduce
1. Copy content from [the following Gist](https://gist.github.com/code-chris/982dd128e5fe3eb58adaac5da947e0a1).
1. Call `editor.setData()` with the content from step 1.
1. Repeat step 2.

## Current result
The editor crashes.

## Error
```
fastdiff.js:176 Uncaught TypeError: arr.slice is not a function
    at cutAndReverse (fastdiff.js:176)
    at findChangeBoundaryIndexes (fastdiff.js:131)
    at Function.fastDiff (fastdiff.js:99)
    at diff (diff.js:37)
    at Renderer._diffNodeLists (renderer.js:584)
    at Renderer._updateChildrenMappings (renderer.js:264)
    at Renderer.render (renderer.js:176)
    at View._render (view.js:667)
    at View.on (view.js:182)
    at View.fire (emittermixin.js:204)
```

## GIF
![bug_cke5](https://user-images.githubusercontent.com/20988892/52621452-9b370080-2ea7-11e9-8930-02add8e60611.gif)