@Reinmar 2019-02-12
> Yes and no :D Previously `render()` took around 23 seconds for the sample data, now it takes around 200 - 300ms. However, the other parts of the pipeline still takes visible amount of time

It's understandable and acceptable (at this stage) when loading a bigger blob of data takes that much time. We should be focused on the part which took the majority of that time and that's `diff()`. Going down from 23s to 200ms is the biggest win we could have – going down another 50ms will be really hard now.

> which means end to end tests for larger data sets still takes up to few seconds to be completed.

I don't understand. Can't we use a data set of a size which will give now results <100ms and would take 10s before?

The point of an e2e test is to make sure the problem's never back. That it doesn't appear in some other place. The problem was that for a data set of a reasonable size the `setData()` call took 20s. It's down to 200ms. Let's make sure it stays there.