@Mgsy 2019-02-12
>https://github.com/ckeditor/ckeditor5-utils/pull/274#issuecomment-462662831

Now I can't reproduce it. Interesting, it would be nice if you could take a look at it 🤔