@f1ames 2019-02-12
> https://github.com/ckeditor/ckeditor5-utils/pull/274#issuecomment-462662831

@Mgsy I will take a look. Is it possible that you had some cached / not up-to-date version? The https://github.com/ckeditor/ckeditor5-utils/pull/274/commits/2263658374b3fb47c33f263581bb44929ab07b39 commit changed how non-array / array like elements are detected and converted into arrays so it could affected the issue you have mentioned.