@Mgsy 2019-02-12
>@Mgsy I will take a look. Is it possible that you had some cached / not up-to-date version?

I suppose it could be this. Now it works fine for me.