@f1ames 2019-02-12
> I don't understand. Can't we use a data set of a size which will give now results <100ms and would take 10s before?
>
> The point of an e2e test is to make sure the problem's never back. That it doesn't appear in some other place. The problem was that for a data set of a reasonable size the `setData()` call took 20s. It's down to 200ms. Let's make sure it stays there.

What I meant earlier is that for data which took around 10s before, `render()` took ~8s and rest of the pipeline took ~2s. Now for such cases `render()` time is probably ~100ms but the rest of the pipeline still takes ~2s. So it is not possible to create e2e test which now takes 100ms and before the changes took 10s (because of entire pipeline additional processing time). And for smaller data sets the difference is not that visible (in the overall time), something like 500ms (with `fastDiff`) vs 700ms (without `fastDiff`) on Chrome (and doubles on Firefox 😞).