@f1ames 2019-02-13
I have added rendering performance tests in https://github.com/ckeditor/ckeditor5-engine/pull/1676.